pyproject.toml
[project]
name = "umi-to-lerobot"
version = "0.1.0"
description = "Convert UMI datasets to LeRobot Dataset v3 format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "zarr<3",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "pillow>=10.0.0",
    "lerobot>=0.4.2",
    "imagecodecs>=2026.1.1",
]
